* deleted unneeded files * clearing out most of warnings

+++ drivers/net/phy/phy.c
@@ -773,7 +773,7 @@ int phy_reset(struct phy_device *phydev)
 	}
 #endif
 
-	printf("phy write %x %x\n", phydev, devad);
+	printf("phy write %x %x\n", (uint32_t)phydev, devad);
 	if (phy_write(phydev, devad, MII_BMCR, BMCR_RESET) < 0) {
 		debug("PHY reset failed\n");
 		return -1;
